refactor: consul, move SD struct from config into consul discovery package
