+++ discovery/consul/consul.go
@@ -25,12 +25,14 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	consul "github.com/hashicorp/consul/api"
-	"github.com/mwitkow/go-conntrack"
+	conntrack "github.com/mwitkow/go-conntrack"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/config"
+	"github.com/prometheus/prometheus/pkg/targetgroup"
+	configUtil "github.com/prometheus/prometheus/util/config"
 	"github.com/prometheus/prometheus/util/httputil"
 	"github.com/prometheus/prometheus/util/strutil"
+	yamlUtil "github.com/prometheus/prometheus/util/yaml"
 )
 
 const (
@@ -75,7 +77,48 @@ var (
 		},
 		[]string{"endpoint", "call"},
 	)
+
+	// DefaultSDConfig is the default Consul SD configuration.
+	DefaultSDConfig = SDConfig{
+		TagSeparator: ",",
+		Scheme:       "http",
+	}
 )
+
+// SDConfig is the configuration for Consul service discovery.
+type SDConfig struct {
+	Server       string            `yaml:"server"`
+	Token        configUtil.Secret `yaml:"token,omitempty"`
+	Datacenter   string            `yaml:"datacenter,omitempty"`
+	TagSeparator string            `yaml:"tag_separator,omitempty"`
+	Scheme       string            `yaml:"scheme,omitempty"`
+	Username     string            `yaml:"username,omitempty"`
+	Password     configUtil.Secret `yaml:"password,omitempty"`
+	// The list of services for which targets are discovered.
+	// Defaults to all services if empty.
+	Services []string `yaml:"services"`
+
+	TLSConfig configUtil.TLSConfig `yaml:"tls_config,omitempty"`
+	// Catches all undefined fields and must be empty after parsing.
+	XXX map[string]interface{} `yaml:",inline"`
+}
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface.
+func (c *SDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	*c = DefaultSDConfig
+	type plain SDConfig
+	err := unmarshal((*plain)(c))
+	if err != nil {
+		return err
+	}
+	if err := yamlUtil.CheckOverflow(c.XXX, "consul_sd_config"); err != nil {
+		return err
+	}
+	if strings.TrimSpace(c.Server) == "" {
+		return fmt.Errorf("Consul SD configuration requires a server address")
+	}
+	return nil
+}
 
 func init() {
 	prometheus.MustRegister(rpcFailuresCount)
@@ -98,7 +141,7 @@ type Discovery struct {
 }
 
 // NewDiscovery returns a new Discovery for the given config.
-func NewDiscovery(conf *config.ConsulSDConfig, logger log.Logger) (*Discovery, error) {
+func NewDiscovery(conf *SDConfig, logger log.Logger) (*Discovery, error) {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
@@ -160,7 +203,7 @@ func (d *Discovery) shouldWatch(name string) bool {
 }
 
 // Run implements the TargetProvider interface.
-func (d *Discovery) Run(ctx context.Context, ch chan<- []*config.TargetGroup) {
+func (d *Discovery) Run(ctx context.Context, ch chan<- []*targetgroup.Group) {
 	// Watched services and their cancelation functions.
 	services := map[string]func(){}
 
@@ -243,7 +286,7 @@ func (d *Discovery) Run(ctx context.Context, ch chan<- []*config.TargetGroup) {
 				select {
 				case <-ctx.Done():
 					return
-				case ch <- []*config.TargetGroup{{Source: name}}:
+				case ch <- []*targetgroup.Group{{Source: name}}:
 				}
 			}
 		}
@@ -259,7 +302,7 @@ type consulService struct {
 	logger       log.Logger
 }
 
-func (srv *consulService) watch(ctx context.Context, ch chan<- []*config.TargetGroup) {
+func (srv *consulService) watch(ctx context.Context, ch chan<- []*targetgroup.Group) {
 	catalog := srv.client.Catalog()
 
 	lastIndex := uint64(0)
@@ -291,7 +334,7 @@ func (srv *consulService) watch(ctx context.Context, ch chan<- []*config.TargetG
 		}
 		lastIndex = meta.LastIndex
 
-		tgroup := config.TargetGroup{
+		tgroup := targetgroup.Group{
 			Source:  srv.name,
 			Labels:  srv.labels,
 			Targets: make([]model.LabelSet, 0, len(nodes)),
@@ -339,7 +382,7 @@ func (srv *consulService) watch(ctx context.Context, ch chan<- []*config.TargetG
 		select {
 		case <-ctx.Done():
 			return
-		case ch <- []*config.TargetGroup{&tgroup}:
+		case ch <- []*targetgroup.Group{&tgroup}:
 		}
 	}
 }
